refactor(chat): code review fixes and issue picker dismiss bug
- Move parseUserMentions to shared TabView/utils/ location
- Add task-mention test cases (7 new tests)
- Make file mentions clickable in MessageList (parity with UserMessage)
- Remove useMemo overhead in LinkedTaskChip
- Fix || to ?? in LinkedIssuePill for consistent nullish coalescing
- Remove redundant type checks in already-filtered arrays
- Fix issue picker dialog dismissing on mouse click by making it modal
- Prevent dropdown focus restoration racing with dialog via onCloseAutoFocus
