+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatMastraPane/ChatMastraInterface/components/ChatMastraMessageList/components/UserMessage/utils/parseUserMentions/index.ts
@@ -1,4 +1,4 @@
 export {
 	parseUserMentions,
 	type UserMentionSegment,
-} from "./parseUserMentions";
+} from "../../../../../../../../utils/parseUserMentions";
