+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatPane/ChatInterface/components/IssueLinkCommand/IssueLinkCommand.tsx
@@ -112,6 +112,7 @@ export function IssueLinkCommand({
 		<CommandDialog
 			open={open}
 			onOpenChange={handleOpenChange}
+			modal
 			title="Link issue"
 			description="Search for an issue to link"
 			showCloseButton={false}
