+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatPane/ChatInterface/components/MessageList/MessageList.tsx
@@ -11,9 +11,10 @@ import { FileIcon, FileTextIcon, ImageIcon } from "lucide-react";
 import { useCallback } from "react";
 import { HiMiniChatBubbleLeftRight } from "react-icons/hi2";
 import { useTabsStore } from "renderer/stores/tabs/store";
-import { parseUserMentions } from "../../../../ChatMastraPane/ChatMastraInterface/components/ChatMastraMessageList/components/UserMessage/utils/parseUserMentions";
 import { LinkedTaskChip } from "../../../../components/LinkedTaskChip";
+import { parseUserMentions } from "../../../../utils/parseUserMentions";
 import type { InterruptedMessagePreview } from "../../types";
+import { normalizeWorkspaceFilePath } from "../../utils/file-paths";
 import { MessagePartsRenderer } from "../MessagePartsRenderer";
 import { MessageScrollbackRail } from "./components/MessageScrollbackRail";
 
@@ -96,11 +97,11 @@ export function MessageList({
 								.map((p) => p.text)
 								.join("");
 							const fileParts = msg.parts.filter((p) => p.type === "file");
-							const imageParts = fileParts.filter(
-								(p) => p.type === "file" && p.mediaType.startsWith("image/"),
+							const imageParts = fileParts.filter((p) =>
+								p.mediaType.startsWith("image/"),
 							);
 							const nonImageParts = fileParts.filter(
-								(p) => p.type === "file" && !p.mediaType.startsWith("image/"),
+								(p) => !p.mediaType.startsWith("image/"),
 							);
 
 							const mentionSegments = textContent
@@ -127,48 +128,42 @@ export function MessageList({
 								>
 									{imageParts.length > 0 && (
 										<div className="flex max-w-[85%] flex-wrap gap-2">
-											{imageParts.map((p) =>
-												p.type === "file" ? (
-													<button
-														key={`${msg.id}-img-${p.url}`}
-														type="button"
-														className="cursor-zoom-in"
-														onClick={() => handleImageClick(p.url)}
-													>
-														<img
-															src={p.url}
-															alt={p.filename || "Attached image"}
-															className="max-h-48 rounded-lg object-contain"
-														/>
-													</button>
-												) : null,
-											)}
+											{imageParts.map((p) => (
+												<button
+													key={`${msg.id}-img-${p.url}`}
+													type="button"
+													className="cursor-zoom-in"
+													onClick={() => handleImageClick(p.url)}
+												>
+													<img
+														src={p.url}
+														alt={p.filename || "Attached image"}
+														className="max-h-48 rounded-lg object-contain"
+													/>
+												</button>
+											))}
 										</div>
 									)}
 									{nonImageParts.length > 0 && (
 										<div className="flex max-w-[85%] flex-wrap gap-1.5">
-											{nonImageParts.map((p) =>
-												p.type === "file" ? (
-													<FileChip
-														key={`${msg.id}-file-${p.url}`}
-														filename={p.filename || ""}
-														mediaType={p.mediaType}
-													/>
-												) : null,
-											)}
+											{nonImageParts.map((p) => (
+												<FileChip
+													key={`${msg.id}-file-${p.url}`}
+													filename={p.filename || ""}
+													mediaType={p.mediaType}
+												/>
+											))}
 										</div>
 									)}
 									{taskMentions.length > 0 && workspaceId && (
 										<div className="flex max-w-[85%] flex-wrap justify-end gap-2">
-											{taskMentions.map((segment, segIdx) =>
-												segment.type === "task-mention" ? (
-													<LinkedTaskChip
-														key={`${msg.id}-task-${segIdx}`}
-														slug={segment.slug}
-														workspaceId={workspaceId}
-													/>
-												) : null,
-											)}
+											{taskMentions.map((segment, segIdx) => (
+												<LinkedTaskChip
+													key={`${msg.id}-task-${segIdx}`}
+													slug={segment.slug}
+													workspaceId={workspaceId}
+												/>
+											))}
 										</div>
 									)}
 									{hasNonTaskContent && (
@@ -185,10 +180,27 @@ export function MessageList({
 													);
 												}
 												if (segment.type === "file-mention") {
+													const normalizedPath = normalizeWorkspaceFilePath({
+														filePath: segment.relativePath,
+														workspaceRoot: workspaceCwd,
+													});
+													const canOpen =
+														Boolean(normalizedPath) && Boolean(workspaceId);
+
 													return (
-														<span
+														<button
+															type="button"
 															key={`${msg.id}-seg-${segIdx}`}
-															className="mx-0.5 inline-flex items-center gap-0.5 rounded-md bg-primary/15 px-1.5 py-0.5 font-mono text-xs text-primary"
+															className="mx-0.5 inline-flex items-center gap-0.5 rounded-md bg-primary/15 px-1.5 py-0.5 font-mono text-xs text-primary transition-colors hover:bg-primary/22 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary/40 disabled:cursor-default disabled:opacity-60"
+															onClick={() => {
+																if (!normalizedPath || !workspaceId) return;
+																addFileViewerPane(workspaceId, {
+																	filePath: normalizedPath,
+																	isPinned: true,
+																});
+															}}
+															disabled={!canOpen}
+															aria-label={`Open file ${segment.relativePath}`}
 														>
 															<span className="font-semibold text-primary">
 																@
@@ -196,7 +208,7 @@ export function MessageList({
 															<span className="text-primary/95">
 																{segment.relativePath}
 															</span>
-														</span>
+														</button>
 													);
 												}
 												return null;
