+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/utils/parseUserMentions/index.ts
@@ -1,0 +1,4 @@
+export {
+	parseUserMentions,
+	type UserMentionSegment,
+} from "./parseUserMentions";
